requirements.txt
# 深度学习框架
torch>=2.1.0
torchvision>=0.16.0

# Transformers和模型
transformers>=4.35.0
accelerate>=0.24.0
safetensors>=0.4.0

# 图像处理
pillow>=10.0.0
opencv-python>=4.8.0

# 数值计算
numpy>=1.24.0

# 音频处理（用于ImageBind）
librosa>=0.10.0
soundfile>=0.12.0

# 配置和工具
pyyaml>=6.0.0

# 机器学习指标
scikit-learn>=1.3.0

# 其他
sentencepiece>=0.1.99
protobuf>=3.20.0

